pykt: Finish moving dispatch to plugins
