+++ pykt/plugins.py
@@ -2,6 +2,7 @@ import io
 import os
 import ast
 import functools
+import re
 import textwrap
 
 from tempfile import NamedTemporaryFile
@@ -90,63 +91,54 @@ class KotlinTranspilerPlugins:
 
     def visit_range(self, node, vargs: List[str]) -> str:
         if len(node.args) == 1:
-            return "(0..{})".format(vargs[0])
+            return "(0..{}-1)".format(vargs[0])
         elif len(node.args) == 2:
-            return "({}..{})".format(vargs[0], vargs[1])
+            return "({}..{}-1)".format(vargs[0], vargs[1])
         elif len(node.args) == 3:
-            return "({}..{}).step_by({})".format(vargs[0], vargs[1], vargs[2])
+            return "({}..{}-1 step {})".format(vargs[0], vargs[1], vargs[2])
 
         raise Exception(
             "encountered range() call with unknown parameters: range({})".format(vargs)
         )
 
     def visit_print(self, node, vargs: List[str]) -> str:
-        placeholders = []
-        for n in node.args:
-            placeholders.append("{}")
-        return 'println!("{0}",{1});'.format(" ".join(placeholders), ", ".join(vargs))
+        def _format(arg):
+            if arg.isdigit():
+                return arg
+            if re.match(r"'.*'", arg) or re.match(r'".*"', arg):
+                return arg[1:-1]
+            else:
+                return f"${arg}"
+
+        vargs_str = " ".join([f"{_format(arg)}" for arg in vargs])
+        return f'println("{vargs_str}")'
 
     def visit_min_max(self, node, vargs, is_max: bool) -> str:
-        self._usings.add("std::cmp")
         min_max = "max" if is_max else "min"
+        self._usings.add(f"kotlin.math.{min_max}")
         self._typename_from_annotation(node.args[0])
         if hasattr(node.args[0], "container_type"):
-            node.result_type = True
-            return f"{vargs[0]}.iter().{min_max}()"
+            return f"maxOf({vargs[0]})"
         else:
             all_vargs = ", ".join(vargs)
-            return f"cmp::{min_max}({all_vargs})"
+            return f"{min_max}({all_vargs})"
 
-    @staticmethod
-    def visit_cast(node, vargs, cast_to: str) -> str:
-        return f"{vargs[0]} as {cast_to}"
-
-    @staticmethod
-    def visit_asyncio_run(node, vargs) -> str:
-        return f"block_on({vargs[0]})"
+    def visit_floor(self, node, vargs) -> str:
+        self._usings.add("kotlin.math.floor")
+        return f"floor({vargs[0]}).toInt()"
 
 
 # small one liners are inlined here as lambdas
 SMALL_DISPATCH_MAP = {
-    "str": lambda n, vargs: f"&{vargs[0]}.to_string()",
-    "len": lambda n, vargs: f"{vargs[0]}.len()",
-    "enumerate": lambda n, vargs: f"{vargs[0]}.iter().enumerate()",
-    "sum": lambda n, vargs: f"{vargs[0]}.iter().sum()",
-    "int": functools.partial(KotlinTranspilerPlugins.visit_cast, cast_to="i32"),
+    "str": lambda n, vargs: f"{vargs[0]}.toString()",
+    # TODO: strings use .length
+    "len": lambda n, vargs: f"{vargs[0]}.size",
+    "int": lambda n, vargs: f"{vargs[0]}.toInt()",
     "bool": lambda n, vargs: f"({vargs[0]} != 0)",
-    "float": functools.partial(KotlinTranspilerPlugins.visit_cast, cast_to="f64"),
-    # as usize below is a hack to pass comb_sort.rs. Need a better solution
-    "floor": lambda n, vargs: f"{vargs[0]}.floor() as usize",
-    "reversed": lambda n, vargs: f"{vargs[0]}.iter().rev()",
-    "map": lambda n, vargs: f"{vargs[1]}.iter().map({vargs[0]})",
-    "filter": lambda n, vargs: f"{vargs[1]}.into_iter().filter({vargs[0]})",
-    "list": lambda n, vargs: f"{vargs[0]}.collect::<Vec<_>>()",
-    "asyncio.run": KotlinTranspilerPlugins.visit_asyncio_run,
+    "reversed": lambda n, vargs: f"{vargs[0]}.reversed()",
 }
 
-SMALL_USINGS_MAP = {
-    "asyncio.run": "futures::executor::block_on",
-}
+SMALL_USINGS_MAP = {}
 
 DISPATCH_MAP = {
     "max": functools.partial(KotlinTranspilerPlugins.visit_min_max, is_max=True),
@@ -154,6 +146,7 @@ DISPATCH_MAP = {
     "range": KotlinTranspilerPlugins.visit_range,
     "xrange": KotlinTranspilerPlugins.visit_range,
     "print": KotlinTranspilerPlugins.visit_print,
+    "floor": KotlinTranspilerPlugins.visit_floor,
 }
 
 MODULE_DISPATCH_TABLE = {}
